103 Glan Road
At Waterloo we were fortunate in catching a train for Leatherhead, where we hired a trap at the
station inn and drove for four or five miles through the lovely Surrey lanes. It was a perfect day,
with a bright sun and a few fleecy clouds in the heavens. The trees and wayside hedges were just
throwing out their first green shoots, and the air was full of the pleasant smell of the moist
earth. To me at least there was a strange contrast between the sweet promise of the spring and this
sinister quest upon which we were engaged. My companion sat in the front of the trap, his arms
folded, his hat pulled down over his eyes, and his chin sunk upon his breast, buried in the deepest
thought. Suddenly, however, he started, tapped me on the shoulder, and pointed over the meadows.

"Look there!" said he.