12 Valentia Road
It was, indeed, our friend the financier. I was shocked by the change which had come over him, for
his face which was naturally of a broad and massive mould, was now pinched and fallen in, while his
hair seemed to me at least a shade whiter. He entered with a weariness and lethargy which was even
more painful than his violence of the morning before, and he dropped heavily into the armchair which
I pushed forward for him.

"I do not know what I have done to be so severely tried," said he. "Only two days ago I was a happy
and prosperous man, without a care in the world. Now I am left to a lonely and dishonoured age. One
sorrow comes close upon the heels of another. My niece, Mary, has deserted me."

"Deserted you?"

"Yes. Her bed this morning had not been slept in, her room was empty, and a note for me lay upon the
hall table. I had said to her last night, in sorrow and not in anger, that if she had married my boy
all might have been well with him. Perhaps it was thoughtless of me to say so. It is to that remark
that she refers in this note: