141 Sunningdale Drive
"He didn't mind me; in fact, he took a fancy to me, for at the time when he saw me first I was a
youngster of twelve or so. This would be in the year 1878, after he had been eight or nine years in
England. He begged my father to let me live with him and he was very kind to me in his way. When he
was sober he used to be fond of playing backgammon and draughts with me, and he would make me his
representative both with the servants and with the tradespeople, so that by the time that I was
sixteen I was quite master of the house. I kept all the keys and could go where I liked and do what
I liked, so long as I did not disturb him in his privacy. There was one singular exception, however,
for he had a single room, a lumber-room up among the attics, which was invariably locked, and which
he would never permit either me or anyone else to enter. With a boy's curiosity I have peeped
through the keyhole, but I was never able to see more than such a collection of old trunks and
bundles as would be expected in such a room.

"One day--it was in March, 1883--a letter with a foreign stamp lay upon the table in front of the
colonel's plate. It was not a common thing for him to receive letters, for his bills were all paid
in ready money, and he had no friends of any sort. 'From India!' said he as he took it up,
'Pondicherry postmark! What can this be?' Opening it hurriedly, out there jumped five little dried
orange pips, which pattered down upon his plate. I began to laugh at this, but the laugh was struck
from my lips at the sight of his face. His lip had fallen, his eyes were protruding, his skin the
colour of putty, and he glared at the envelope which he still held in his trembling hand, 'K. K.
K.!' he shrieked, and then, 'My God, my God, my sins have overtaken me!'

"'What is it, uncle?' I cried.

"'Death,' said he, and rising from the table he retired to his room, leaving me palpitating with
horror. I took up the envelope and saw scrawled in red ink upon the inner flap, just above the gum,
the letter K three times repeated. There was nothing else save the five dried pips. What could be
the reason of his overpowering terror? I left the breakfast-table, and as I ascended the stair I met
him coming down with an old rusty key, which must have belonged to the attic, in one hand, and a
small brass box, like a cashbox, in the other.

"'They may do what they like, but I'll checkmate them still,' said he with an oath. 'Tell Mary that
I shall want a fire in my room to-day, and send down to Fordham, the Horsham lawyer.'