18 Harvey Street
"Why did you come away to consult me in such a hurry?" asked Sherlock Holmes, with his finger-tips
together and his eyes to the ceiling.

Again a startled look came over the somewhat vacuous face of Miss Mary Sutherland. "Yes, I did bang
out of the house," she said, "for it made me angry to see the easy way in which Mr. Windibank--that
is, my father--took it all. He would not go to the police, and he would not go to you, and so at
last, as he would do nothing and kept on saying that there was no harm done, it made me mad, and I
just on with my things and came right away to you."