67 Harvey Street
"The fish that you have tattooed immediately above your right wrist could only have been done in
China. I have made a small study of tattoo marks and have even contributed to the literature of the
subject. That trick of staining the fishes' scales of a delicate pink is quite peculiar to China.
When, in addition, I see a Chinese coin hanging from your watch-chain, the matter becomes even more
simple."

Mr. Jabez Wilson laughed heavily. "Well, I never!" said he. "I thought at first that you had done
something clever, but I see that there was nothing in it, after all."

"I begin to think, Watson," said Holmes, "that I make a mistake in explaining. 'Omne ignotum pro
magnifico,' you know, and my poor little reputation, such as it is, will suffer shipwreck if I am so
candid. Can you not find the advertisement, Mr. Wilson?"

"Yes, I have got it now," he answered with his thick red finger planted halfway down the column.
"Here it is. This is what began it all. You just read it for yourself, sir."

I took the paper from him and read as follows:

"To the Red-headed League: On account of the bequest of the late Ezekiah Hopkins, of Lebanon,
Pennsylvania, U. S. A., there is now another vacancy open which entitles a member of the League to a
salary of £4 a week for purely nominal services. All red-headed men who are sound in body and mind
and above the age of twenty-one years, are eligible. Apply in person on Monday, at eleven o'clock,
to Duncan Ross, at the offices of the League, 7 Pope's Court, Fleet Street."