29 Wellswood Gardens
"My dear Holmes!"

"I have notes of several similar cases, though none, as I remarked before, which were quite as
prompt. My whole examination served to turn my conjecture into a certainty. Circumstantial evidence
is occasionally very convincing, as when you find a trout in the milk, to quote Thoreau's example."

"But I have heard all that you have heard."

"Without, however, the knowledge of pre-existing cases which serves me so well. There was a parallel
instance in Aberdeen some years back, and something on very much the same lines at Munich the year
after the Franco-Prussian War. It is one of these cases--but, hullo, here is Lestrade! Good-
afternoon, Lestrade! You will find an extra tumbler upon the sideboard, and there are cigars in the
box."

The official detective was attired in a pea-jacket and cravat, which gave him a decidedly nautical
appearance, and he carried a black canvas bag in his hand. With a short greeting he seated himself
and lit the cigar which had been offered to him.

"What's up, then?" asked Holmes with a twinkle in his eye. "You look dissatisfied."