16 Glendale
"Now a pawnbroker's business is mostly done of an evening, Mr. Holmes, especially Thursday and
Friday evening, which is just before pay-day; so it would suit me very well to earn a little in the
mornings. Besides, I knew that my assistant was a good man, and that he would see to anything that
turned up.

"'That would suit me very well,' said I. 'And the pay?'

"'Is £4 a week.'

"'And the work?'

"'Is purely nominal.'

"'What do you call purely nominal?'