19 Wellswood Gardens
"Subtle enough and horrible enough. When a doctor does go wrong he is the first of criminals. He has
nerve and he has knowledge. Palmer and Pritchard were among the heads of their profession. This man
strikes even deeper, but I think, Watson, that we shall be able to strike deeper still. But we shall
have horrors enough before the night is over; for goodness' sake let us have a quiet pipe and turn
our minds for a few hours to something more cheerful."

About nine o'clock the light among the trees was extinguished, and all was dark in the direction of
the Manor House. Two hours passed slowly away, and then, suddenly, just at the stroke of eleven, a
single bright light shone out right in front of us.