75 Sunningdale Drive
"What on earth has that to do with it?" I ejaculated.

"My dear Watson, you as a medical man are continually gaining light as to the tendencies of a child
by the study of the parents. Don't you see that the converse is equally valid. I have frequently
gained my first real insight into the character of parents by studying their children. This child's
disposition is abnormally cruel, merely for cruelty's sake, and whether he derives this from his
smiling father, as I should suspect, or from his mother, it bodes evil for the poor girl who is in
their power."

"I am sure that you are right, Mr. Holmes," cried our client. "A thousand things come back to me
which make me certain that you have hit it. Oh, let us lose not an instant in bringing help to this
poor creature."

"We must be circumspect, for we are dealing with a very cunning man. We can do nothing until seven
o'clock. At that hour we shall be with you, and it will not be long before we solve the mystery."

We were as good as our word, for it was just seven when we reached the Copper Beeches, having put up
our trap at a wayside public-house. The group of trees, with their dark leaves shining like
burnished metal in the light of the setting sun, were sufficient to mark the house even had Miss
Hunter not been standing smiling on the door-step.