87 Beaconside
"'Tell me, dad,' said she, looking, I thought, a little disturbed, 'did you give Lucy, the maid,
leave to go out to-night?'

"'Certainly not.'

"'She came in just now by the back door. I have no doubt that she has only been to the side gate to
see someone, but I think that it is hardly safe and should be stopped.'

"'You must speak to her in the morning, or I will if you prefer it. Are you sure that everything is
fastened?'