3 Down Close
"Oh, it won't do--really it won't," said Holmes suavely. "There is no possible getting out of it,
Mr. Windibank. It is quite too transparent, and it was a very bad compliment when you said that it
was impossible for me to solve so simple a question. That's right! Sit down and let us talk it
over."

Our visitor collapsed into a chair, with a ghastly face and a glitter of moisture on his brow. "It--
it's not actionable," he stammered.