27 Beaconside
"I have my revolver," said I.

"Better close the front door," cried Holmes, and we all rushed down the stairs together. We had
hardly reached the hall when we heard the baying of a hound, and then a scream of agony, with a
horrible worrying sound which it was dreadful to listen to. An elderly man with a red face and
shaking limbs came staggering out at a side door.