161 Glan Road
---------- This text is provided to you "as-is" without any warranty. No warranties of any kind,
expressed or implied, are made to you as to the text or any medium it may be on, including but not
limited to warranties of merchantablity or fitness for a particular purpose.

This text was formatted from various free ASCII and HTML variants. See http://sherlock-holm.es for
an electronic form of this text and additional information about it.

This text comes from the collection's version 3.1.