90 Down Close
"When did she vanish, then?"

"At the wedding breakfast."

"Indeed. This is more interesting than it promised to be; quite dramatic, in fact."

"Yes; it struck me as being a little out of the common."

"They often vanish before the ceremony, and occasionally during the honeymoon; but I cannot call to
mind anything quite so prompt as this. Pray let me have the details."

"I warn you that they are very incomplete."