23 Wellswood Gardens
"Were you engaged to the gentleman at this time?"

"Oh, yes, Mr. Holmes. We were engaged after the first walk that we took. Hosmer--Mr. Angel--was a
cashier in an office in Leadenhall Street--and--"

"What office?"

"That's the worst of it, Mr. Holmes, I don't know."

"Where did he live, then?"