58 Tamworth Drive
That is all."

"Terse and to the point," remarked Holmes, stretching his long, thin legs towards the fire.

"There was a paragraph amplifying this in one of the society papers of the same week. Ah, here it
is:

"'There will soon be a call for protection in the marriage market, for the present free-trade
principle appears to tell heavily against our home product. One by one the management of the noble
houses of Great Britain is passing into the hands of our fair cousins from across the Atlantic. An
important addition has been made during the last week to the list of the prizes which have been
borne away by these charming invaders. Lord St. Simon, who has shown himself for over twenty years
proof against the little god's arrows, has now definitely announced his approaching marriage with
Miss Hatty Doran, the fascinating daughter of a California millionaire. Miss Doran, whose graceful
figure and striking face attracted much attention at the Westbury House festivities, is an only
child, and it is currently reported that her dowry will run to considerably over the six figures,
with expectancies for the future. As it is an open secret that the Duke of Balmoral has been
compelled to sell his pictures within the last few years, and as Lord St. Simon has no property of
his own save the small estate of Birchmoor, it is obvious that the Californian heiress is not the
only gainer by an alliance which will enable her to make the easy and common transition from a
Republican lady to a British peeress.'"

"Anything else?" asked Holmes, yawning.

"Oh, yes; plenty. Then there is another note in the Morning Post to say that the marriage would be
an absolutely quiet one, that it would be at St. George's, Hanover Square, that only half a dozen
intimate friends would be invited, and that the party would return to the furnished house at
Lancaster Gate which has been taken by Mr. Aloysius Doran. Two days later--that is, on Wednesday
last--there is a curt announcement that the wedding had taken place, and that the honeymoon would be
passed at Lord Backwater's place, near Petersfield. Those are all the notices which appeared before
the disappearance of the bride."