4 Down Close
"'Arthur!' I screamed, 'you villain! you thief! How dare you touch that coronet?'

"The gas was half up, as I had left it, and my unhappy boy, dressed only in his shirt and trousers,
was standing beside the light, holding the coronet in his hands. He appeared to be wrenching at it,
or bending it with all his strength. At my cry he dropped it from his grasp and turned as pale as
death. I snatched it up and examined it. One of the gold corners, with three of the beryls in it,
was missing.