112 Holmefield Avenue
"'Perhaps you will have the kindness to wait in this room for a few minutes,' said he, throwing open
another door. It was a quiet, little, plainly furnished room, with a round table in the centre, on
which several German books were scattered. Colonel Stark laid down the lamp on the top of a
harmonium beside the door. 'I shall not keep you waiting an instant,' said he, and vanished into the
darkness.

"I glanced at the books upon the table, and in spite of my ignorance of German I could see that two
of them were treatises on science, the others being volumes of poetry. Then I walked across to the
window, hoping that I might catch some glimpse of the country-side, but an oak shutter, heavily
barred, was folded across it. It was a wonderfully silent house. There was an old clock ticking
loudly somewhere in the passage, but otherwise everything was deadly still. A vague feeling of
uneasiness began to steal over me. Who were these German people, and what were they doing living in
this strange, out-of-the-way place? And where was the place? I was ten miles or so from Eyford, that
was all I knew, but whether north, south, east, or west I had no idea. For that matter, Reading, and
possibly other large towns, were within that radius, so the place might not be so secluded, after
all. Yet it was quite certain, from the absolute stillness, that we were in the country. I paced up
and down the room, humming a tune under my breath to keep up my spirits and feeling that I was
thoroughly earning my fifty-guinea fee.

"Suddenly, without any preliminary sound in the midst of the utter stillness, the door of my room
swung slowly open. The woman was standing in the aperture, the darkness of the hall behind her, the
yellow light from my lamp beating upon her eager and beautiful face. I could see at a glance that
she was sick with fear, and the sight sent a chill to my own heart. She held up one shaking finger
to warn me to be silent, and she shot a few whispered words of broken English at me, her eyes
glancing back, like those of a frightened horse, into the gloom behind her.