71 Down Close
"You may address me as the Count Von Kramm, a Bohemian nobleman. I understand that this gentleman,
your friend, is a man of honour and discretion, whom I may trust with a matter of the most extreme
importance. If not, I should much prefer to communicate with you alone."

I rose to go, but Holmes caught me by the wrist and pushed me back into my chair. "It is both, or
none," said he. "You may say before this gentleman anything which you may say to me."

The Count shrugged his broad shoulders. "Then I must begin," said he, "by binding you both to
absolute secrecy for two years; at the end of that time the matter will be of no importance. At
present it is not too much to say that it is of such weight it may have an influence upon European
history."

"I promise," said Holmes.