11 Sunningdale Drive
"This aroused my curiosity, so when I went out for a walk in the grounds with my charge, I strolled
round to the side from which I could see the windows of this part of the house. There were four of
them in a row, three of which were simply dirty, while the fourth was shuttered up. They were
evidently all deserted. As I strolled up and down, glancing at them occasionally, Mr. Rucastle came
out to me, looking as merry and jovial as ever.

"'Ah!' said he, 'you must not think me rude if I passed you without a word, my dear young lady. I
was preoccupied with business matters.'