58 Balcombe Close
"It may be so, or it may not, Mr. Holmes," said he, "but if you are so very sharp you ought to be
sharp enough to know that it is you who are breaking the law now, and not me. I have done nothing
actionable from the first, but as long as you keep that door locked you lay yourself open to an
action for assault and illegal constraint."

"The law cannot, as you say, touch you," said Holmes, unlocking and throwing open the door, "yet
there never was a man who deserved punishment more. If the young lady has a brother or a friend, he
ought to lay a whip across your shoulders. By Jove!" he continued, flushing up at the sight of the
bitter sneer upon the man's face, "it is not part of my duties to my client, but here's a hunting
crop handy, and I think I shall just treat myself to--" He took two swift steps to the whip, but
before he could grasp it there was a wild clatter of steps upon the stairs, the heavy hall door
banged, and from the window we could see Mr. James Windibank running at the top of his speed down
the road.

"There's a cold-blooded scoundrel!" said Holmes, laughing, as he threw himself down into his chair
once more. "That fellow will rise from crime to crime until he does something very bad, and ends on
a gallows. The case has, in some respects, been not entirely devoid of interest."

"I cannot now entirely see all the steps of your reasoning," I remarked.