3 Badgers Dene
"Exactly so. But I am now sleeping in the middle one."

"Pending the alterations, as I understand. By the way, there does not seem to be any very pressing
need for repairs at that end wall."

"There were none. I believe that it was an excuse to move me from my room."