76 Down Close
"This hat is three years old. These flat brims curled at the edge came in then. It is a hat of the
very best quality. Look at the band of ribbed silk and the excellent lining. If this man could
afford to buy so expensive a hat three years ago, and has had no hat since, then he has assuredly
gone down in the world."

"Well, that is clear enough, certainly. But how about the foresight and the moral retrogression?"

Sherlock Holmes laughed. "Here is the foresight," said he putting his finger upon the little disc
and loop of the hat-securer.  "They are never sold upon hats. If this man ordered one, it is a sign
of a certain amount of foresight, since he went out of his way to take this precaution against the
wind. But since we see that he has broken the elastic and has not troubled to replace it, it is
obvious that he has less foresight now than formerly, which is a distinct proof of a weakening
nature. On the other hand, he has endeavoured to conceal some of these stains upon the felt by
daubing them with ink, which is a sign that he has not entirely lost his self-respect."

"Your reasoning is certainly plausible."