109 Down Close
There is ever a flaw, however, in the best laid of human plans, and the murderers of John Openshaw
were never to receive the orange pips which would show them that another, as cunning and as resolute
as themselves, was upon their track. Very long and very severe were the equinoctial gales that year.
We waited long for news of the Lone Star of Savannah, but none ever reached us. We did at last hear
that somewhere far out in the Atlantic a shattered stern-post of a boat was seen swinging in the
trough of a wave, with the letters "L. S." carved upon it, and that is all which we shall ever know
of the fate of the Lone Star.

Isa Whitney, brother of the late Elias Whitney, D.D., Principal of the Theological College of St.
George's, was much addicted to opium. The habit grew upon him, as I understand, from some foolish
freak when he was at college; for having read De Quincey's description of his dreams and sensations,
he had drenched his tobacco with laudanum in an attempt to produce the same effects. He found, as so
many more have done, that the practice is easier to attain than to get rid of, and for many years he
continued to be a slave to the drug, an object of mingled horror and pity to his friends and
relatives. I can see him now, with yellow, pasty face, drooping lids, and pin-point pupils, all
huddled in a chair, the wreck and ruin of a noble man.

One night--it was in June, '89--there came a ring to my bell, about the hour when a man gives his
first yawn and glances at the clock. I sat up in my chair, and my wife laid her needle-work down in
her lap and made a little face of disappointment.