146 Wantage Close
It was close upon four before the door opened, and a drunken-looking groom, ill-kempt and side-
whiskered, with an inflamed face and disreputable clothes, walked into the room. Accustomed as I was
to my friend's amazing powers in the use of disguises, I had to look three times before I was
certain that it was indeed he. With a nod he vanished into the bedroom, whence he emerged in five
minutes tweed-suited and respectable, as of old. Putting his hands into his pockets, he stretched
out his legs in front of the fire and laughed heartily for some minutes.

"Well, really!" he cried, and then he choked and laughed again until he was obliged to lie back,
limp and helpless, in the chair.

"What is it?"

"It's quite too funny. I am sure you could never guess how I employed my morning, or what I ended by
doing."

"I can't imagine. I suppose that you have been watching the habits, and perhaps the house, of Miss
Irene Adler."

"Quite so; but the sequel was rather unusual. I will tell you, however. I left the house a little
after eight o'clock this morning in the character of a groom out of work. There is a wonderful
sympathy and freemasonry among horsey men. Be one of them, and you will know all that there is to
know. I soon found Briony Lodge. It is a bijou villa, with a garden at the back, but built out in
front right up to the road, two stories. Chubb lock to the door. Large sitting-room on the right
side, well furnished, with long windows almost to the floor, and those preposterous English window
fasteners which a child could open. Behind there was nothing remarkable, save that the passage
window could be reached from the top of the coach-house. I walked round it and examined it closely
from every point of view, but without noting anything else of interest.