10 Balcombe Close
"I called about that beggarman, Boone--the one who was charged with being concerned in the
disappearance of Mr. Neville St. Clair, of Lee."

"Yes. He was brought up and remanded for further inquiries."

"So I heard. You have him here?"

"In the cells."

"Is he quiet?"

"Oh, he gives no trouble. But he is a dirty scoundrel."