169 Down Close
"Holmes," I said, "you have drawn a net round this man from which he cannot escape, and you have
saved an innocent human life as truly as if you had cut the cord which was hanging him. I see the
direction in which all this points. The culprit is--"

"Mr. John Turner," cried the hotel waiter, opening the door of our sitting-room, and ushering in a
visitor.