21 Badgers Dene
Holmes laughed softly to himself and stretched himself out upon the cushioned seat. "Both you and
the coroner have been at some pains," said he, "to single out the very strongest points in the young
man's favour. Don't you see that you alternately give him credit for having too much imagination and
too little? Too little, if he could not invent a cause of quarrel which would give him the sympathy
of the jury; too much, if he evolved from his own inner consciousness anything so outré as a dying
reference to a rat, and the incident of the vanishing cloth. No, sir, I shall approach this case
from the point of view that what this young man says is true, and we shall see whither that
hypothesis will lead us. And now here is my pocket Petrarch, and not another word shall I say of
this case until we are on the scene of action. We lunch at Swindon, and I see that we shall be there
in twenty minutes."

It was nearly four o'clock when we at last, after passing through the beautiful Stroud Valley, and
over the broad gleaming Severn, found ourselves at the pretty little country-town of Ross. A lean,
ferret-like man, furtive and sly-looking, was waiting for us upon the platform. In spite of the
light brown dustcoat and leather-leggings which he wore in deference to his rustic surroundings, I
had no difficulty in recognising Lestrade, of Scotland Yard. With him we drove to the Hereford Arms
where a room had already been engaged for us.