104 Moor End
"'My dear Mr. Sherlock Holmes: "'Lord Backwater tells me that I may place implicit reliance upon
your judgment and discretion. I have determined, therefore, to call upon you and to consult you in
reference to the very painful event which has occurred in connection with my wedding. Mr. Lestrade,
of Scotland Yard, is acting already in the matter, but he assures me that he sees no objection to
your co-operation, and that he even thinks that it might be of some assistance. I will call at four
o'clock in the afternoon, and, should you have any other engagement at that time, I hope that you
will postpone it, as this matter is of paramount importance. "'Yours faithfully, "'St. Simon.'

"It is dated from Grosvenor Mansions, written with a quill pen, and the noble lord has had the
misfortune to get a smear of ink upon the outer side of his right little finger," remarked Holmes as
he folded up the epistle.

"He says four o'clock. It is three now. He will be here in an hour."