42 Beaconside
"I cannot imagine."

"No, I don't think you would guess. Every pocket stuffed with pennies and half-pennies--421 pennies
and 270 half-pennies. It was no wonder that it had not been swept away by the tide. But a human body
is a different matter. There is a fierce eddy between the wharf and the house. It seemed likely
enough that the weighted coat had remained when the stripped body had been sucked away into the
river."

"But I understand that all the other clothes were found in the room. Would the body be dressed in a
coat alone?"

"No, sir, but the facts might be met speciously enough. Suppose that this man Boone had thrust
Neville St. Clair through the window, there is no human eye which could have seen the deed. What
would he do then? It would of course instantly strike him that he must get rid of the tell-tale
garments. He would seize the coat, then, and be in the act of throwing it out, when it would occur
to him that it would swim and not sink. He has little time, for he has heard the scuffle downstairs
when the wife tried to force her way up, and perhaps he has already heard from his Lascar
confederate that the police are hurrying up the street. There is not an instant to be lost. He
rushes to some secret hoard, where he has accumulated the fruits of his beggary, and he stuffs all
the coins upon which he can lay his hands into the pockets to make sure of the coat's sinking. He
throws it out, and would have done the same with the other garments had not he heard the rush of
steps below, and only just had time to close the window when the police appeared."

"It certainly sounds feasible."

"Well, we will take it as a working hypothesis for want of a better. Boone, as I have told you, was
arrested and taken to the station, but it could not be shown that there had ever before been
anything against him. He had for years been known as a professional beggar, but his life appeared to
have been a very quiet and innocent one. There the matter stands at present, and the questions which
have to be solved--what Neville St. Clair was doing in the opium den, what happened to him when
there, where is he now, and what Hugh Boone had to do with his disappearance--are all as far from a
solution as ever. I confess that I cannot recall any case within my experience which looked at the
first glance so simple and yet which presented such difficulties."