20 Sunningdale Drive
"So they have. And many men have been wrongfully hanged."

"What is the young man's own account of the matter?"

"It is, I am afraid, not very encouraging to his supporters, though there are one or two points in
it which are suggestive. You will find it here, and may read it for yourself."