122 Down Close
"It was only a joke at first," groaned our visitor. "We never thought that she would have been so
carried away."

"Very likely not. However that may be, the young lady was very decidedly carried away, and, having
quite made up her mind that her stepfather was in France, the suspicion of treachery never for an
instant entered her mind. She was flattered by the gentleman's attentions, and the effect was
increased by the loudly expressed admiration of her mother. Then Mr. Angel began to call, for it was
obvious that the matter should be pushed as far as it would go if a real effect were to be produced.
There were meetings, and an engagement, which would finally secure the girl's affections from
turning towards anyone else. But the deception could not be kept up forever. These pretended
journeys to France were rather cumbrous. The thing to do was clearly to bring the business to an end
in such a dramatic manner that it would leave a permanent impression upon the young lady's mind and
prevent her from looking upon any other suitor for some time to come. Hence those vows of fidelity
exacted upon a Testament, and hence also the allusions to a possibility of something happening on
the very morning of the wedding. James Windibank wished Miss Sutherland to be so bound to Hosmer
Angel, and so uncertain as to his fate, that for ten years to come, at any rate, she would not
listen to another man. As far as the church door he brought her, and then, as he could go no
farther, he conveniently vanished away by the old trick of stepping in at one door of a four-wheeler
and out at the other. I think that was the chain of events, Mr. Windibank!"

Our visitor had recovered something of his assurance while Holmes had been talking, and he rose from
his chair now with a cold sneer upon his pale face.

"It may be so, or it may not, Mr. Holmes," said he, "but if you are so very sharp you ought to be
sharp enough to know that it is you who are breaking the law now, and not me. I have done nothing
actionable from the first, but as long as you keep that door locked you lay yourself open to an
action for assault and illegal constraint."

"The law cannot, as you say, touch you," said Holmes, unlocking and throwing open the door, "yet
there never was a man who deserved punishment more. If the young lady has a brother or a friend, he
ought to lay a whip across your shoulders. By Jove!" he continued, flushing up at the sight of the
bitter sneer upon the man's face, "it is not part of my duties to my client, but here's a hunting
crop handy, and I think I shall just treat myself to--" He took two swift steps to the whip, but
before he could grasp it there was a wild clatter of steps upon the stairs, the heavy hall door
banged, and from the window we could see Mr. James Windibank running at the top of his speed down
the road.