110 Down Close
"'What is it, uncle?' I cried.

"'Death,' said he, and rising from the table he retired to his room, leaving me palpitating with
horror. I took up the envelope and saw scrawled in red ink upon the inner flap, just above the gum,
the letter K three times repeated. There was nothing else save the five dried pips. What could be
the reason of his overpowering terror? I left the breakfast-table, and as I ascended the stair I met
him coming down with an old rusty key, which must have belonged to the attic, in one hand, and a
small brass box, like a cashbox, in the other.

"'They may do what they like, but I'll checkmate them still,' said he with an oath. 'Tell Mary that
I shall want a fire in my room to-day, and send down to Fordham, the Horsham lawyer.'

"I did as he ordered, and when the lawyer arrived I was asked to step up to the room. The fire was
burning brightly, and in the grate there was a mass of black, fluffy ashes, as of burned paper,
while the brass box stood open and empty beside it. As I glanced at the box I noticed, with a start,
that upon the lid was printed the treble K which I had read in the morning upon the envelope.

"'I wish you, John,' said my uncle, 'to witness my will. I leave my estate, with all its advantages
and all its disadvantages, to my brother, your father, whence it will, no doubt, descend to you. If
you can enjoy it in peace, well and good! If you find you cannot, take my advice, my boy, and leave
it to your deadliest enemy. I am sorry to give you such a two-edged thing, but I can't say what turn
things are going to take. Kindly sign the paper where Mr. Fordham shows you.'

"I signed the paper as directed, and the lawyer took it away with him. The singular incident made,
as you may think, the deepest impression upon me, and I pondered over it and turned it every way in
my mind without being able to make anything of it. Yet I could not shake off the vague feeling of
dread which it left behind, though the sensation grew less keen as the weeks passed and nothing
happened to disturb the usual routine of our lives. I could see a change in my uncle, however. He
drank more than ever, and he was less inclined for any sort of society. Most of his time he would
spend in his room, with the door locked upon the inside, but sometimes he would emerge in a sort of
drunken frenzy and would burst out of the house and tear about the garden with a revolver in his
hand, screaming out that he was afraid of no man, and that he was not to be cooped up, like a sheep
in a pen, by man or devil. When these hot fits were over, however, he would rush tumultuously in at
the door and lock and bar it behind him, like a man who can brazen it out no longer against the
terror which lies at the roots of his soul. At such times I have seen his face, even on a cold day,
glisten with moisture, as though it were new raised from a basin.