73 Corbet Close
"And brought Miss Hunter down from London in order to get rid of the disagreeable persistence of Mr.
Fowler."

"That was it, sir."

"But Mr. Fowler being a persevering man, as a good seaman should be, blockaded the house, and having
met you succeeded by certain arguments, metallic or otherwise, in convincing you that your interests
were the same as his."

"Mr. Fowler was a very kind-spoken, free-handed gentleman," said Mrs. Toller serenely.