63 Wantage Close
"Farintosh," said he. "Ah yes, I recall the case; it was concerned with an opal tiara. I think it
was before your time, Watson. I can only say, madam, that I shall be happy to devote the same care
to your case as I did to that of your friend. As to reward, my profession is its own reward; but you
are at liberty to defray whatever expenses I may be put to, at the time which suits you best. And
now I beg that you will lay before us everything that may help us in forming an opinion upon the
matter."

"Alas!" replied our visitor, "the very horror of my situation lies in the fact that my fears are so
vague, and my suspicions depend so entirely upon small points, which might seem trivial to another,
that even he to whom of all others I have a right to look for help and advice looks upon all that I
tell him about it as the fancies of a nervous woman. He does not say so, but I can read it from his
soothing answers and averted eyes. But I have heard, Mr. Holmes, that you can see deeply into the
manifold wickedness of the human heart. You may advise me how to walk amid the dangers which
encompass me."

"I am all attention, madam."

"My name is Helen Stoner, and I am living with my stepfather, who is the last survivor of one of the
oldest Saxon families in England, the Roylotts of Stoke Moran, on the western border of Surrey."

Holmes nodded his head. "The name is familiar to me," said he.