29 Balcombe Close
"Well, I got the two dozen from a salesman in Covent Garden."

"Indeed? I know some of them. Which was it?"

"Breckinridge is his name."

"Ah! I don't know him. Well, here's your good health landlord, and prosperity to your house. Good-
night."

"Now for Mr. Breckinridge," he continued, buttoning up his coat as we came out into the frosty air.
"Remember, Watson that though we have so homely a thing as a goose at one end of this chain, we have
at the other a man who will certainly get seven years' penal servitude unless we can establish his
innocence. It is possible that our inquiry may but confirm his guilt; but, in any case, we have a
line of investigation which has been missed by the police, and which a singular chance has placed in
our hands. Let us follow it out to the bitter end. Faces to the south, then, and quick march!"