137 Bowler Street
"But you can know nothing of this?"

"Excuse me, I know everything of it. You are endeavouring to trace some geese which were sold by
Mrs. Oakshott, of Brixton Road, to a salesman named Breckinridge, by him in turn to Mr. Windigate,
of the Alpha, and by him to his club, of which Mr. Henry Baker is a member."