79 Badgers Dene
"Well, it is very ingenious," said I, laughing; "but since, as you said just now, there has been no
crime committed, and no harm done save the loss of a goose, all this seems to be rather a waste of
energy."

Sherlock Holmes had opened his mouth to reply, when the door flew open, and Peterson, the
commissionaire, rushed into the apartment with flushed cheeks and the face of a man who is dazed
with astonishment.

"The goose, Mr. Holmes! The goose, sir!" he gasped.

"Eh? What of it, then? Has it returned to life and flapped off through the kitchen window?" Holmes
twisted himself round upon the sofa to get a fairer view of the man's excited face.