132 Sunningdale Drive
"Pray do so."

"Well, now, in considering this case there are two points about young McCarthy's narrative which
struck us both instantly, although they impressed me in his favour and you against him. One was the
fact that his father should, according to his account, cry 'Cooee!' before seeing him. The other was
his singular dying reference to a rat. He mumbled several words, you understand, but that was all
that caught the son's ear. Now from this double point our research must commence, and we will begin
it by presuming that what the lad says is absolutely true."

"What of this 'Cooee!' then?"

"Well, obviously it could not have been meant for the son. The son, as far as he knew, was in
Bristol. It was mere chance that he was within earshot. The 'Cooee!' was meant to attract the
attention of whoever it was that he had the appointment with. But 'Cooee' is a distinctly Australian
cry, and one which is used between Australians. There is a strong presumption that the person whom
McCarthy expected to meet him at Boscombe Pool was someone who had been in Australia."