56 Wantage Close
"They considered that it might be caused by Arthur's closing his bedroom door."

"A likely story! As if a man bent on felony would slam his door so as to wake a household. What did
they say, then, of the disappearance of these gems?"

"They are still sounding the planking and probing the furniture in the hope of finding them."

"Have they thought of looking outside the house?"