49 Wantage Close
I waited until midnight, but there was no sign of his return, so I retired to my room. It was no
uncommon thing for him to be away for days and nights on end when he was hot upon a scent, so that
his lateness caused me no surprise. I do not know at what hour he came in, but when I came down to
breakfast in the morning there he was with a cup of coffee in one hand and the paper in the other,
as fresh and trim as possible.

"You will excuse my beginning without you, Watson," said he, "but you remember that our client has
rather an early appointment this morning."