90 Sunningdale Drive
"To do nothing whatever. There will probably be some small unpleasantness. Do not join in it. It
will end in my being conveyed into the house. Four or five minutes afterwards the sitting-room
window will open. You are to station yourself close to that open window."

"Yes."

"You are to watch me, for I will be visible to you."

"Yes."

"And when I raise my hand--so--you will throw into the room what I give you to throw, and will, at
the same time, raise the cry of fire. You quite follow me?"