62 Larch Walk
"Your own little income," he asked, "does it come out of the business?"

"Oh, no, sir. It is quite separate and was left me by my uncle Ned in Auckland. It is in New Zealand
stock, paying 4½ per cent. Two thousand five hundred pounds was the amount, but I can only touch the
interest."

"You interest me extremely," said Holmes. "And since you draw so large a sum as a hundred a year,
with what you earn into the bargain, you no doubt travel a little and indulge yourself in every way.
I believe that a single lady can get on very nicely upon an income of about £60."

"I could do with much less than that, Mr. Holmes, but you understand that as long as I live at home
I don't wish to be a burden to them, and so they have the use of the money just while I am staying
with them. Of course, that is only just for the time. Mr. Windibank draws my interest every quarter
and pays it over to mother, and I find that I can do pretty well with what I earn at typewriting. It
brings me twopence a sheet, and I can often do from fifteen to twenty sheets in a day."